rotate container runtime logs
